Benchmark factoring.   - Separate basics, strings, and collections tests.   - Add Result.Guid column for separate null/Distinct coverage.
