+++ csharp/BSOA/BSOA.Benchmarks/NonBsoaModel/Generator.cs
@@ -41,7 +41,8 @@ namespace BSOA.Benchmarks.NonBsoaModel
             for (int i = 0; i < ResultCount; ++i)
             {
                 Result result = new Result();
-                result.RuleId = null; // $"SCAN{r.Next(RuleCount - 2):D4}";
+                // result.Guid = null;
+                result.RuleId = $"SCAN{r.Next(RuleCount - 2):D4}";
                 result.IsActive = (i % 3 == 0);
                 result.BaselineState = (i % 10 == 4 ? BaselineState.New : BaselineState.Unchanged);
                 result.StartLine = r.Next(1000);
