+++ csharp/BSOA/BSOA.Benchmarks/NonBsoaModel/Result.cs
@@ -5,6 +5,7 @@ namespace BSOA.Benchmarks.NonBsoaModel
 {
     public class Result
     {
+        public string Guid { get; set; }
         public string RuleId { get; set; }
         public bool IsActive { get; set; }
         public string Message { get; set; }
